learning erlang and elixir
I really need to come up with a nicer solution for selectively
disabling tab indentation.  Or maybe it would be better to keep it off
by default and only enable it for certain languages?  Given how often
I accidentally mixed tabs and spaces in the past, that might be the
wiser course of action.