Genius.com's search engine sucks so simplify the output name format
This is good for single songs but it would be good to have different
naming conventions for channels and playlists, (eg, downloading a
whole album or discography), but there's still the problem of getting
search terms for lyrics.  Perhaps that would be best solved by
grepping the output of ffprobe or opuscomment.